Automate example_app directory renaming + provide version value in /status endpoint
